+++ src/webapp/src/components/pipelines/Transforms.vue
@@ -6,28 +6,33 @@ export default {
   name: 'Transforms',
   data() {
     return {
-      selectedTransformOption: null,
+      selectedTransformOption: null
     }
   },
   methods: {
     saveTransformAndGoToSchedules() {
-      this.$store.dispatch('configuration/updateRecentELTSelections', {
-        type: 'transform',
-        value: this.selectedTransformOption
-      }).then(() => {
-        this.$router.push({ name: 'schedules' })
-        Vue.toasted.global.success(`Transform Saved - ${this.selectedTransformOption.label}`)
-      })
+      this.$store
+        .dispatch('configuration/updateRecentELTSelections', {
+          type: 'transform',
+          value: this.selectedTransformOption
+        })
+        .then(() => {
+          this.$router.push({ name: 'schedules' })
+          Vue.toasted.global.success(
+            `Transform Saved - ${this.selectedTransformOption.label}`
+          )
+        })
     },
     updateTransformTypeSelection(transformType) {
       this.selectedTransformOption = transformType
     }
   },
   computed: {
-    ...mapState('configuration', ['recentELTSelections', 'transformOptions']),
+    ...mapState('configuration', ['recentELTSelections', 'transformOptions'])
   },
   created() {
-    const defaultTransform = this.recentELTSelections.transform || this.transformOptions[0]
+    const defaultTransform =
+      this.recentELTSelections.transform || this.transformOptions[0]
     this.selectedTransformOption = defaultTransform
   }
 }
@@ -67,49 +72,66 @@ export default {
                 </a>
               </div>
               <div class="message-body">
-                <p>Currently, for the UI, we only support the <em>default transforms</em> that ship with Meltano. The CLI however <a href="https://www.meltano.com/docs/tutorial.html#advanced-adding-custom-transformations-and-models">provides more options</a>. Longer term we ancipate this view to support transform creation, editing, and selection.</p>
+                <p>
+                  Currently, for the UI, we only support the
+                  <em>default transforms</em> that ship with Meltano. The CLI
+                  however
+                  <a
+                    href="https://www.meltano.com/docs/tutorial.html#advanced-adding-custom-transformations-and-models"
+                    >provides more options</a
+                  >. Longer term we ancipate this view to support transform
+                  creation, editing, and selection.
+                </p>
                 <p>Current Options:</p>
                 <ul>
-                  <li>Run (ELT): run the default transforms with extract and load</li>
-                  <li>Skip (EL): do not run the default transforms with extract and load</li>
-                  <li>Only (T): only run default transforms, do not extract and load</li>
+                  <li>
+                    Run (ELT): run the default transforms with extract and load
+                  </li>
+                  <li>
+                    Skip (EL): do not run the default transforms with extract
+                    and load
+                  </li>
+                  <li>
+                    Only (T): only run default transforms, do not extract and
+                    load
+                  </li>
                 </ul>
               </div>
             </article>
           </div>
 
-          <hr>
+          <hr />
 
           <div class="level">
-              <div class="level-left">
-                <div class="buttons has-addons">
-                  <button
-                    v-for='transformOption in transformOptions'
-                    :key='transformOption.label'
-                    class="button is-outlined"
-                    :class="{
-                      'is-selected is-interactive-secondary': transformOption === selectedTransformOption
-                    }"
-                    @click="updateTransformTypeSelection(transformOption)"
-                  >
-                    {{ transformOption.label }}
-                  </button>
-                </div>
+            <div class="level-left">
+              <div class="buttons has-addons">
+                <button
+                  v-for="transformOption in transformOptions"
+                  :key="transformOption.label"
+                  class="button is-outlined"
+                  :class="{
+                    'is-selected is-interactive-secondary':
+                      transformOption === selectedTransformOption
+                  }"
+                  @click="updateTransformTypeSelection(transformOption)"
+                >
+                  {{ transformOption.label }}
+                </button>
               </div>
+            </div>
 
-              <div class="level-right">
-                <a
-                  class="button is-interactive-primary"
-                  @click='saveTransformAndGoToSchedules'>Save</a>
-              </div>
-
+            <div class="level-right">
+              <a
+                class="button is-interactive-primary"
+                @click="saveTransformAndGoToSchedules"
+                >Save</a
+              >
+            </div>
           </div>
         </div>
       </div>
     </div>
-
   </div>
 </template>
 
-<style lang="scss">
-</style>
+<style lang="scss"></style>