+++ src/webapp/src/router/index.js
@@ -5,7 +5,6 @@ import AnalyzeSettings from '@/components/analyze/AnalyzeSettings'
 import CreateScheduleModal from '@/components/pipelines/CreateScheduleModal'
 import Design from '@/components/analyze/Design'
 import EntitiesSelectorModal from '@/components/pipelines/EntitiesSelectorModal'
-import Entities from '@/components/pipelines/Entities'
 import Extractors from '@/components/pipelines/Extractors'
 import ExtractorSettingsModal from '@/components/pipelines/ExtractorSettingsModal'
 import Loaders from '@/components/pipelines/Loaders'
@@ -49,7 +48,7 @@ const router = new Router({
           }
         },
         {
-          path: 'extractors/:extractor',
+          path: 'extractors/:extractor/configure',
           name: 'extractorSettings',
           components: {
             default: Extractors,
@@ -60,20 +59,10 @@ const router = new Router({
           }
         },
         {
-          path: 'entities',
-          name: 'entities',
-          components: {
-            default: Entities
-          },
-          meta: {
-            isModal: false
-          }
-        },
-        {
-          path: 'entities/:extractor',
+          path: 'extractors/:extractor/select',
           name: 'extractorEntities',
           components: {
-            default: Entities,
+            default: Extractors,
             extractorEntities: EntitiesSelectorModal
           },
           meta: {